Support dragging a tab around the screen
A tab starts travelling when it's grabbed with the mouse and
then the pointer moves out of the tab area without releasing.
The travelling image is a copy of how the tab looks when selected,
without the closing X. When the tab lands back on the tab area,
we swap appropriately. If released outside the tab area, for now
it does nothing. Moving out to a new window will be added later.